feat(mnestix-related-infos-footer): fix footer layout, remove duplicated infos
